Fix the card components, add some styling to inspection maker
