+++ pages/admin/makeInspection.tsx
@@ -1,0 +1,15 @@
+import { CssBaseline } from "@mui/material";
+import InspectionMaker from "../../src/components/inspections/Reports/inspectionMaker";
+
+const makeInspection = () => {
+  return (
+    <>
+      <main>
+        <CssBaseline />
+        <InspectionMaker />
+      </main>
+    </>
+  );
+};
+
+export default makeInspection;
