+++ src/components/inspections/Card/uncompletedCard.tsx
@@ -7,19 +7,20 @@ import {
   CardContent,
   Typography,
 } from "@mui/material";
-import Link from "next/link";
 import ArrowForwardIcon from "@mui/icons-material/ArrowForward";
-import { UncompletedInspectionCardProps } from "../../../../ts/interfaces/roomReport.interfaces";
+import { InspectionCardProps } from "../../../../ts/interfaces/roomReport.interfaces";
 import { useDispatch } from "react-redux";
 import { setInspectionSelectionData } from "../../../../slices/inspectionSelectionSlice";
+import { useRouter } from "next/router";
 
 const UncompletedCard = ({
   card_id,
   inspectionProps,
 }: {
   card_id: number;
-  inspectionProps: UncompletedInspectionCardProps;
+  inspectionProps: InspectionCardProps;
 }) => {
+  const router = useRouter();
   const dispatch = useDispatch();
   const handleClicked = () => {
     dispatch(
@@ -28,35 +29,38 @@ const UncompletedCard = ({
         inspections: inspectionProps,
       })
     );
+    router.push("/admin/makeInspection");
   };
 
   return (
-    <Card>
-      <Box sx={{ display: "inline-flex", width: "fill", alignItems: "center" }}>
-        <CardContent sx={{ width: "180px", mr: 0 }}>
-          <Typography variant="h6">
-            <b>Room{" " + inspectionProps.room_name} </b>
-          </Typography>
-          <Typography variant="h6">
-            Floor {inspectionProps.floor_number} in{" "}
-            {inspectionProps.building_name} Building
-          </Typography>
-          <Typography>
-            Cleaned by{" "}
-            {inspectionProps.team_members
-              .map((member) => member.first_name + " " + member.last_name)
-              .join(", ")}
-          </Typography>
-        </CardContent>
-
-        <CardActions sx={{ ml: 0 }}>
-          <Link href={{ pathname: "inspectionMaker" }}>
-            <Button endIcon={<ArrowForwardIcon />} onClick={handleClicked}>
-              Inspect
-            </Button>
-          </Link>
-        </CardActions>
-      </Box>
+    <Card sx={{ display: "flex", alignItems: "center" }}>
+      <CardContent
+        sx={{
+          display: "flex",
+          flexDirection: "column",
+          minWidth: 0,
+          flex: 1,
+        }}
+      >
+        <Typography variant="h6">
+          <b>Room{" " + inspectionProps.room_name} </b>
+        </Typography>
+        <Typography noWrap variant="h6">
+          Floor {inspectionProps.floor_number} in{" "}
+          {inspectionProps.building_name} Building
+        </Typography>
+        <Typography noWrap>
+          Cleaned by{" "}
+          {inspectionProps.team_members
+            .map((member) => member.first_name + " " + member.last_name)
+            .join(", ")}
+        </Typography>
+      </CardContent>
+      <CardActions sx={{ ml: 0, minWidth: "fit-content" }}>
+        <Button endIcon={<ArrowForwardIcon />} onClick={handleClicked}>
+          Inspect
+        </Button>
+      </CardActions>
     </Card>
   );
 };
