+++ src/components/inspections/Grid/inspectionCardGrid.tsx
@@ -3,7 +3,10 @@ import { Inspect_Status } from "@prisma/client";
 import { Inspection } from "../../../../ts/types/db.interfaces";
 import CompletedCard from "../Card/completedCard";
 import UncompletedCard from "../Card/uncompletedCard";
-import { toUncompletedInspectionCardProps } from "../../../../ts/interfaces/roomReport.interfaces";
+import {
+  toCompletedInspectionCardProps,
+  toUncompletedInspectionCardProps,
+} from "../../../../ts/interfaces/roomReport.interfaces";
 import { useSelector } from "react-redux";
 import {
   getInspectedReports,
@@ -13,20 +16,19 @@ import {
 const InspectionCardGrid = ({ status }: { status: Inspect_Status }) => {
   const inspected = useSelector(getInspectedReports);
   const notInspected = useSelector(getNotInspectedReports);
+  console.log(inspected);
 
   return (
-    <>
-      <Grid
-        container
-        item
-        direction="row"
-        xs={4}
-        rowSpacing={2}
-        sx={{ width: "max-content", overflow: "visible", p: 2 }}
-      >
-        {CardCondition({ status, inspected, notInspected })}
-      </Grid>
-    </>
+    <Grid
+      container
+      direction="row"
+      spacing={2}
+      p={2}
+      maxHeight={"75vh"} // This is needed to allow the scrollbar to scroll to the bottom (basically 100vh - other stuff)
+      overflow={"auto"}
+    >
+      {CardCondition({ status, inspected, notInspected })}
+    </Grid>
   );
 };
 
@@ -42,17 +44,21 @@ const CardCondition = ({
   switch (status) {
     case Inspect_Status.INSPECTED:
       return inspected.map((inspection, index) => (
-        <UncompletedCard
-          card_id={index}
-          inspectionProps={toUncompletedInspectionCardProps(inspection)}
-        />
+        <Grid item xs={12} md={6} xl={4}>
+          <CompletedCard
+            card_id={index}
+            inspectionProps={toCompletedInspectionCardProps(inspection)}
+          />
+        </Grid>
       ));
     case Inspect_Status.NOT_INSPECTED:
       return notInspected.map((inspection, index) => (
-        <UncompletedCard
-          card_id={index}
-          inspectionProps={toUncompletedInspectionCardProps(inspection)}
-        />
+        <Grid item xs={12} md={6} xl={4}>
+          <UncompletedCard
+            card_id={index}
+            inspectionProps={toUncompletedInspectionCardProps(inspection)}
+          />
+        </Grid>
       ));
     default:
       return <></>;
